# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "okten",
    "version": "0.1",
    "description": "test doc",
    "contact": {
      "name": "Yurii",
      "email": "yurii@gmail.com"
    }
  },
  "paths": {
...
    "/comments/user/{userId}": {
      "get": {
        "operationId": "getUserComments",
        "tags": ["Comments"],
        "description": "get user all comments",
        "parameters": [
          {
            "name": "userId",
            "type": "string",
            "in": "path",
            "description": "user id from params",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/commentResponse"
              }
            },
            "description": "all user comments"
          },

          "400": {
            "schema": {
              "$ref": "#/definitions/errorResponse"
            },
            "description": "some error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "commentResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1,
          "description": "id comment"
        },

        "text": {
          "type": "string",
          "example": "something_text",
          "description": "text content"
        },

        "authorId": {
          "type": "number",
          "example": 115,
          "description": "id user"
        },

        "postId": {
          "type": "number",
          "example": 10,
          "description": "id post"
        },

        "createAt": {
          "type": "string",
          "example": "2022-04-17T10:27:43.000Z",
          "description": "date of writing comment"
        },

        "deleteAt": {
          "type": "null",
          "example": null,
          "description": "delete date"
        }
      }
    },
...
    "errorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Some Wrong",
          "description": "Standard error response"
        }
      }
    },
...
  }
}